+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **Version 0.1.2** (unreleased)
  - Fix program version that is displayed when running `oxipng -V`
  - Ensure `--quiet` mode is actually quiet (@SethDusek [#20](https://github.com/shssoichiro/oxipng/pull/20))
+ - Write status/debug information to stderr instead of stdout
  - Use heuristics to determine best combination for `-o1` ([#21](https://github.com/shssoichiro/oxipng/issues/21))
 
 **Version 0.1.1**